more tests and file renamed
